pyproject.toml
[project]
name = "llm-ocr-doc-intel"
version = "0.1.0"
description = "Processes PSEG utility bills using OCR, LangGraph, and LLMs, storing results in PostgreSQL, all orchestrated by Netflix Conductor."
authors = [
    {name = "Frank",email = "frank@example.com"}
]
readme = "README.md"
requires-python = ">=3.10"

[tool.poetry]
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
ollama = "^0.4.8"
pydantic = "^2.5.3"
pydantic-settings = "^2.2.1"
pytesseract = "^0.3.13"
pillow = "^10.0.0"
conductor-python = "^1.0"
pdf2image = "^1.17.0"
requests = "^2.32.3"
fastapi = "^0.115.12"
psycopg2-binary = "^2.9.9"
opencv-python-headless = "^4.9.0.80"
numpy = "^1.26"
python-dotenv = "^1.0.0"
httpx = "^0.28.1"
openai = "^1.79.0"
streamlit = "^1.45.1"
streamlit-pydantic = "^0.6.0"
watchdog = "^6.0.0"
instructor = "^1.2.3"
litellm = "^1.34.30"
uvicorn = {version = "^0.34.2", extras = ["standard"]}
torch = "==2.2.2"
torchvision = "==0.17.2"
torchaudio = "==2.2.2"
docling = "^1.0"
transformers = "^4.38.2"
sentencepiece = "^0.2.0"
llama-parse = "^0.6.23"
llama-index-core = "^0.12.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
# ... existing code ...
